requirements.txt
flask
transformers
torch
newspaper3k
feedparser
lxml #Required parser backend used by newspaper3k
